CRUD, Flyway Initial database, corrected data access services and implement JDBC service

+++ src/main/java/com/jazy/customer/CustomerDao.java
@@ -6,11 +6,11 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> findCustomerById(int id);
+    Optional<Customer> findCustomerById(long id);
     void saveCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    boolean existsCustomerById(int id);
-    void deleteCustomer(int id);
+    boolean existsCustomerById(long id);
+    void deleteCustomer(long id);
     void updateCustomer(Customer customer);
